fix: make the BITS id generator portable to systems without hostname command
The current BITS id generator requires that the hostname command is
available on the machine; however, Windows and some Linux platforms do
not include this command.  To support better portability the Node.js
os.hostname() is used and if that fails the environment variable
HOSTNAME is used.  The hostname command is only used as a final
fallback.

In addition, when hostname command isn't available on the machine it
throws a possibly confusing error, so an additional catch statement has
been added when executing the command.

closes #19