fix(client): add month and day to activity timestamp
The time of day is not enough information for an activity.  Add the month and day.
I don't think there is a reason to add the year.

Also, remove the 'cursor:pointer' directive.  That pointer makes the user
think they can click on the activity for more information.  Since the user
can't do that, don't mislead them.

BITS-17