feat(upgrade): Show upgrade errors in the OMGs UI
Adds an icon to the UI in the OMGs page near the Current Version to indicate if errors occurred in
the most recent upgrade.  Clicking on the icon will show the errors.  Also removes the branch name
(no longer used) from the Current Version string in the upper right corner of the OMGs page.

Closes BITS-109, BITS-110, BITS-111, BITS-112, BITS-113