fix(upgrade): Ensure file removal is complete
Minor debug output changes to Helper.  Added Helper.promiseWhile to enable promise-based while
loops.  Previous rm -rf commands were not erasing all files in the specified directories, so all rm
-rf commands have been converted to _rmrf calls, which use a loop to verify that all files have been
removed.  Note that this _rmrf function includes a watchdog counter to ensure that the loop will
terminate. Fixed a bug in _countFiles which caused an erroneous return if the given path was not
valid on the system.

closes BITS-117