fix: Address a race condition between modules and helpers
Helpers and modules are managed on two separate interfaces.  It is possible that a client gets the
module load complete before it gets the helper loaded message which results in using a helper API
before it has been 'required'.  By moving the onList() to the end of the event queue, the module
loaded message will be processed last and the helper message should be processed first.

BSUM-58